+++ spec/jahuty/snippet_spec.rb
@@ -1,11 +1,11 @@
 module Jahuty
   RSpec.describe Snippet do
-    describe "::get" do
+    describe "::render" do
       before { Jahuty.key = 'kn2Kj5ijmT2pH6ZKqAQyNexUqKeRM4VG6DDgWN1lIcc' }
 
       it "returns content" do
-        expect(Snippet.get(1)).to have_attributes({
-          id: 1, content: "This is my first snippet!"
+        expect(Snippet.render(1)).to have_attributes({
+          content: "This is my first snippet!"
         })
       end
     end